backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
playwright==1.41.0
playwright-stealth==2.0.0
aiohttp==3.13.2
beautifulsoup4==4.12.3
openai==1.12.0
python-multipart==0.0.6
youtube-transcript-api==1.2.4
requests==2.32.5
alembic>=1.13.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
ruff>=0.3.0
python-jose[cryptography]>=3.3.0
resend>=2.0.0
PyYAML>=6.0
stripe>=8.0.0
openpyxl>=3.1.0
